refactor layout /sidebar & hooks
