+++ src/components/Layout.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
 import classnames from 'classnames';
-import { useSidebar } from './sidebarContext';
-import { useLayout } from './layoutContext';
+import { useSidebar } from '../hooks/useSidebar';
+import { useLayout } from '../hooks/useLayout';
 
 export interface LayoutProps extends React.HTMLAttributes<HTMLDivElement> {
   rtl?: boolean;
