+++ src/hooks/useProSidebar.tsx
@@ -1,5 +1,5 @@
-import { useLayout } from '../components/layoutContext';
-import { useSidebar } from '../components/sidebarContext';
+import { useLayout } from './useLayout';
+import { useSidebar } from './useSidebar';
 
 interface ProSidebarFunction {
   /**
